Finish principal board for training fee

+++ app/StoredReportTrait.php
@@ -17,7 +17,7 @@ trait StoredReportTrait
     use FpdiTrait;
     use EmailManagementTrait;
 
-    public function generateFC007($fcLogId, $referenceNumber, $output = true, $processId = null)
+    public function generateFC007($fcLogId, $referenceNumber, $output = true, $processId = null, $principalId = null)
     {
         // initiate fpdi
         $pdf = $this->initiateFpdi('NETI-SGA', 'NETI-SGA', $referenceNumber . ".pdf");
@@ -41,41 +41,66 @@ trait StoredReportTrait
         if ($output == true) {
             $pdf->Output();
         } else {
+            $newStatusId = $this->getNewStatus($processId);
 
-            //delete old file
-            $deleteFile = Storage::disk('public')->delete('F-FC-007/' . $referenceNumber . '.pdf');
+            if ($processId < 4) { //replace old file becuase of signature attachment => update status => send email
+                //---------------------------------------------------------------------------------------//
+                //---------------------------------------------------------------------------------------//
+                //delete old file
+                $deleteFile = Storage::disk('public')->delete('F-FC-007/' . $referenceNumber . '.pdf');
 
-            if (!$deleteFile) {
-                session()->flash('error', 'Sending file saved!');
-            }
+                if (!$deleteFile) {
+                    session()->flash('error', 'Sending file saved!');
+                }
 
-            // save to folder
-            $fileName = $referenceNumber . '.pdf';
-            $filePath = storage_path('app/public/F-FC-007/' . $fileName);
-            $pdfContents = $pdf->Output('', 'S');
-            $storeFile = file_put_contents($filePath, $pdfContents);
+                // save to folder
+                $fileName = $referenceNumber . '.pdf';
+                $filePath = storage_path('app/public/F-FC-007/' . $fileName);
+                $pdfContents = $pdf->Output('', 'S');
+                $storeFile = file_put_contents($filePath, $pdfContents);
 
-            if (!$storeFile) {
-                session()->flash('error', 'Saving file saved!');
-            } else {
+                if (!$storeFile) {
+                    session()->flash('error', 'Saving file saved!');
+                } else {
+                    // save to database
+                    $this->updateStatus($fcLogId, $processId, $newStatusId);
+
+                    // send email notification
+                    $this->sendEmail($newStatusId, $principalId, $processId, $referenceNumber);
+                }
+                //---------------------------------------------------------------------------------------//
+                //---------------------------------------------------------------------------------------//
+            } else { //DONT replace old file  => update status => send email
+                //---------------------------------------------------------------------------------------//
+                //---------------------------------------------------------------------------------------//
                 // save to database
-                $query = Fc007Log::find($fcLogId);
-                $newStatusId = $this->getNewStatus($processId);
-                $update = $this->updateQuery($processId, $query, $newStatusId);
-
-                $this->updateTrait($query, 'sga.process-fc007', $update, "Sending file failed!", "File sent successfully!");
+                $this->updateStatus($fcLogId, $processId, $newStatusId);
 
                 // send email notification
-                $emailData = $this->sendEmailNotification($newStatusId);
-                $subject = $this->fcEmailSubject($newStatusId);
-                foreach ($emailData as $name => $email) {
-                    Mail::to($email)
-                        ->cc('sherwin.roxas@neti.com.ph')
-                        ->send(new SendFc007Email($referenceNumber, $subject, $name));
-                }
-                
-                return $this->redirectRoute('sga.process-fc007', ['processId' => $processId]);
+                $this->sendEmail($newStatusId, $principalId, $processId, $referenceNumber);
+                //---------------------------------------------------------------------------------------//
+                //---------------------------------------------------------------------------------------//
             }
+
+            return $this->redirectRoute('sga.process-fc007', ['processId' => $processId]);
+        }
+    }
+
+    public function updateStatus($fcLogId, $processId, $newStatusId)
+    {
+        $query = Fc007Log::find($fcLogId);
+        $update = $this->updateQuery($processId, $query, $newStatusId);
+        $this->updateTrait($query, 'sga.process-fc007', $update, "Sending file failed!", "File sent successfully!");
+    }
+
+    public function sendEmail($newStatusId, $principalId, $processId, $referenceNumber)
+    {
+        $emailData = $this->sendEmailNotification($newStatusId, $principalId, $processId);
+        $subject = $this->fcEmailSubject($newStatusId);
+        foreach ($emailData as $name => $email) {
+            Mail::to($email)
+                ->cc('sherwin.roxas@neti.com.ph')
+                ->send(new SendFc007Email($referenceNumber, $subject, $name));
         }
     }
 
@@ -94,6 +119,12 @@ trait StoredReportTrait
                     'status_id' => $newStatusId,
                     'approved_by' => Auth::user()->full_name,
                     'approved_at' => now(),
+                ]);
+            case 4:
+                $update = $query->update([
+                    'status_id' => $newStatusId,
+                    'payment_slip_by' => Auth::user()->full_name,
+                    'payment_slip_at' => now(),
                 ]);
                 break;
         }
@@ -115,13 +146,44 @@ trait StoredReportTrait
     public function getNewStatus($processId)
     {
         switch ($processId) {
+            case 1:
+                $statusId = 2;
+                break;
             case 2:
                 $statusId = 3;
                 break;
             case 3:
                 $statusId = 4;
                 break;
+            case 4:
+                $statusId = 5;
+                break;
+            case 5:
+                $statusId = 6;
+                break;
         }
         return $statusId;
+    }
+
+    public function buttonLabel($statusId)
+    {
+        switch ($statusId) {
+            case 1:
+                $label = "Generate";
+                break;
+            case 2:
+                $label = "Verify";
+                break;
+            case 3:
+                $label = "Approve";
+                break;
+            case 4:
+                $label = "Send Payment Slip";
+                break;
+            default:
+                $label = "Upload O.R.";
+                break;
+        }
+        return $label;
     }
 }
